% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcsboost-package.R
\docType{package}
\name{Kcsboost-package}
\alias{Kcsboost}
\alias{Kcsboost-package}
\title{Kcsboost: Instance-dependent cost-sensitive extreme gradient boosting}
\description{
Functions to deal with class imbalance and misclassification costs in multiclass classification algorithms.
}
\keyword{internal}
