% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.K2csboost.R
\name{plot.K2csboost}
\alias{plot.K2csboost}
\title{plot method for K2csboost fits}
\usage{
\method{plot}{K2csboost}(x, legend_position = NULL, \dots)
}
\arguments{
\item{x}{a fitted object of class "\code{K2csboost}".}

\item{legend_position}{string indicating the position of the legend.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
plot expected savings against the iteration index.
}
\details{
\code{plot.K2csboost} plots the evolution of the expected savings. Use of ggplot2 for plotting if available, otherwise basic graphics are used.
}
\examples{
library(Kcsboost)
data(creditcard)
i0 <- which(creditcard$Class == 0)
i1 <- which(creditcard$Class == 1)

set.seed(2020)
i0_train <- sample(i0, size = 0.7 * length(i0))
i1_train <- sample(i1, size = 0.7 * length(i1))

train <- creditcard[ c(i0_train, i1_train), ]
test  <- creditcard[-c(i0_train, i1_train), ]

cost_matrix_train <- cost_matrix_test <- matrix(c(25, 100, 75, 2), ncol = 2, byrow = TRUE)

K2csb<- K2csboost(formula               = Class ~ . - 1,
                  train                 = train,
                  test                  = test,
                  cost_matrix_train     = cost_matrix_train,
                  cost_matrix_test      = cost_matrix_test,
                  nrounds               = 300,
                  early_stopping_rounds = 20,
                  verbose               = 1,
                  print_every_n         = 1)
summary(K2csb)
plot(K2csb, "top")
predict(K2csb, newdata=test)
}
\references{
Coulibaly, K.(2022).\emph{"Statistical modeling of construction probabilities and buildable capacities of parcels based on UrbanSimul geohistorical data"}[Unpublished Master 2 thesis]. INRAE and the University Franche-Comte of Besancon.
}
\seealso{
\code{\link{K2csboost}}, \code{\link{summary.K2csboost}},  \code{\link{predict.K2csboost}}
}
\author{
Kipedene Coulibaly
}
