% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.K2csboost.R
\name{summary.K2csboost}
\alias{summary.K2csboost}
\title{summarizing K2csboost fits}
\usage{
\method{summary}{K2csboost}(object, \dots)
}
\arguments{
\item{object}{An object of class "\code{K2csboost}". A result of a call to the \code{K2csboost} function.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
The function is a method for class \code{K2csboost} objects that summarizes their fit.
}
\details{
\code{summary.K2csboost} prints a summary of the "\code{K2csboost}" object.
}
\examples{
library(Kcsboost)
data(creditcard)
i0 <- which(creditcard$Class == 0)
i1 <- which(creditcard$Class == 1)

set.seed(2020)
i0_train <- sample(i0, size = 0.7 * length(i0))
i1_train <- sample(i1, size = 0.7 * length(i1))

train <- creditcard[ c(i0_train, i1_train), ]
test  <- creditcard[-c(i0_train, i1_train), ]

cost_matrix_train <- cost_matrix_test <- matrix(c(25, 100, 75, 2), ncol = 2, byrow = TRUE)

K2csb<- K2csboost(formula               = Class ~ . - 1,
                  train                 = train,
                  test                  = test,
                  cost_matrix_train     = cost_matrix_train,
                  cost_matrix_test      = cost_matrix_test,
                  nrounds               = 300,
                  early_stopping_rounds = 20,
                  verbose               = 1,
                  print_every_n         = 1)
summary(K2csb)
plot(K2csb, "top")
predict(K2csb, newdata=test)
}
\references{
Coulibaly, K.(2022).\emph{"Statistical modeling of construction probabilities and buildable capacities of parcels based on UrbanSimul geohistorical data"}[Unpublished Master 2 thesis]. INRAE and the University Franche-Comte of Besancon.
}
\seealso{
\code{\link{K2csboost}}, \code{\link{plot.K2csboost}}, \code{\link{predict.K2csboost}}
}
\author{
Kipedene Coulibaly
}
